с

+++ code/datums/components/rotting.dm
@@ -1,5 +1,15 @@
 #define DEAD_TO_ZOMBIE_TIME 5 MINUTES	//Time before death -> raised as zombie (when outside of the city)
-										//(This might not be the exact time)
+										//(This isn't exact time. Extended 5 -> 7 because only takes 2-3 min in testing at 5.)
+
+#define CORPSE_ROT_START_TIME 5 MINUTES
+#define CORPSE_SKELETONIZE_TIME 8 MINUTES
+#define CORPSE_DUST_TIME 10 MINUTES
+
+#define SIMPLE_CORPSE_ROT_START 8 MINUTES
+#define SIMPLE_CORPSE_DUST_TIME 10 MINUTES
+
+#define HUNT_CORPSE_ROT_START 20 MINUTES
+#define HUNT_CORPSE_DUST_TIME 35 MINUTES
 
 /datum/component/rot
 	var/amount = 0
@@ -51,8 +61,9 @@
 		amount -= 5 * time_elapsed
 
 	var/mob/living/carbon/C = parent
+
 	var/is_zombie
-	if(HAS_TRAIT(C, TRAIT_DNR) || isconstruct(C))
+	if(HAS_TRAIT(C, TRAIT_DNR))
 		return
 	if(C.mind)
 		if(C.mind.has_antag_datum(/datum/antagonist/zombie))
@@ -62,13 +73,16 @@
 			qdel(src)
 			return
 
+	// Bodies that ever belonged to a player are exempt from auto-decay/dusting.
+	var/was_player = C.mind || C.last_mind || C.ckey
+	if(was_player && !is_zombie)
+		return
+
 	var/area/A = get_area(C)
 	if (istype(A, /area/rogue/indoors/town))	//Stops rotting inside town buildings; will stop your zombification such as at church or appothocary.
 		return
 	if (istype(A, /area/rogue/indoors/deathsedge))	//Stops rotting inside Death's Edge (Death's Door spell area)
 		return
-
-
 
 	if(!(C.mob_biotypes & (MOB_ORGANIC|MOB_UNDEAD)))
 		qdel(src)
@@ -87,13 +101,13 @@
 	for(var/obj/item/bodypart/B in C.bodyparts)
 		if(!B.skeletonized && B.is_organic_limb())
 			if(!B.rotted)
-				if(amount > 20 MINUTES)
+				if(amount > CORPSE_ROT_START_TIME)
 					B.rotted = TRUE
 					findonerotten = TRUE
 					shouldupdate = TRUE
 					C.apply_status_effect(/datum/status_effect/debuff/rotted_zombie)	//-8 con to rotting zombie corpse.
 			else
-				if(amount > 40 MINUTES)
+				if(amount > CORPSE_SKELETONIZE_TIME)
 					if(!is_zombie)
 						B.skeletonize()
 						if(C.dna && C.dna.species)
@@ -102,15 +116,14 @@
 						shouldupdate = TRUE
 				else
 					findonerotten = TRUE
-		if(amount > 35 MINUTES)  // Code to delete a corpse after 35 minutes if it's not a zombie and not skeletonized. Possible failsafe.
+		if(amount > CORPSE_DUST_TIME)
 			if(!is_zombie)
-				if(!C.client)	// We want to dust NPC bodies, not player bodies.
-					if(B.skeletonized)
-						dustme = TRUE
+				if(B.skeletonized)
+					dustme = TRUE
 
 	if(dustme)
 		qdel(src)
-		return C.dust(drop_items=TRUE)
+		return C.dust(drop_items=FALSE)
 
 	if(findonerotten)
 		var/turf/open/T = C.loc
@@ -133,21 +146,38 @@
 				soundloop.start()
 		C.update_body()
 
+	// Sanity check: if we're a human and we've been buried, we kill the sound.
+	if(ishuman(C))
+		var/mob/living/carbon/human/H = C
+		if(H.buried)
+			soundloop.stop()
+
+/datum/component/rot/simple
+	var/rot_start = SIMPLE_CORPSE_ROT_START
+	var/dust_time = SIMPLE_CORPSE_DUST_TIME
+
 /datum/component/rot/simple/process()
 	..()
 	var/mob/living/L = parent
 	if(L.stat != DEAD)
 		qdel(src)
 		return
-	if(amount > 15 MINUTES)
+	// Player-controlled (or formerly player-controlled) creatures don't auto-decay.
+	if(L.mind || L.ckey)
+		return
+	if(amount > rot_start)
 		if(soundloop && soundloop.stopped)
 			soundloop.start()
 		var/turf/open/T = get_turf(L)
 		if(istype(T))
 			T.pollute_turf(/datum/pollutant/rot, 5)
-	if(amount > 25 MINUTES)
+	if(amount > dust_time)
 		qdel(src)
 		return L.dust(drop_items=TRUE)
+
+/datum/component/rot/simple/hunt
+	rot_start = HUNT_CORPSE_ROT_START
+	dust_time = HUNT_CORPSE_DUST_TIME
 
 /datum/component/rot/gibs
 	amount = MIASMA_GIBS_MOLES
@@ -159,3 +189,10 @@
 	extra_range = 0
 
 #undef DEAD_TO_ZOMBIE_TIME
+#undef CORPSE_ROT_START_TIME
+#undef CORPSE_SKELETONIZE_TIME
+#undef CORPSE_DUST_TIME
+#undef SIMPLE_CORPSE_ROT_START
+#undef SIMPLE_CORPSE_DUST_TIME
+#undef HUNT_CORPSE_ROT_START
+#undef HUNT_CORPSE_DUST_TIME
